+++ src/routes/ConsortiumManager/settings/users/general/PermissionSets/hooks/index.js
@@ -1,0 +1,2 @@
+export { useManageTenantPermissions } from './useManageTenantPermissions';
+export { useSelectPermissionSetById } from './useSelectPermissionSetById';