+++ src/routes/ConsortiumManager/settings/users/general/PermissionSets/hooks/useSelectPermissionSetById.js
@@ -1,0 +1,18 @@
+import { useTenantPermissions } from '../../../../../../../hooks';
+
+export const useSelectPermissionSetById = ({ permissionSetId, tenantId }) => {
+  const { permissions, isLoading } = useTenantPermissions({
+    tenantId,
+    searchParams: {
+      query: 'mutable==true',
+      expandSubs: true,
+    },
+  });
+
+  const selectedPermissionSet = permissions.find(({ id }) => id === permissionSetId);
+
+  return ({
+    isLoading,
+    selectedPermissionSet: selectedPermissionSet || {},
+  });
+};